Reorganized routes folder
