+++ app/clients/dropbox/routes/checkUnsavedAccount.js
@@ -1,0 +1,6 @@
+module.exports = function checkUnsavedAccount(req, res, next) {
+  if (!req.session.unsavedAccount) return next(new Error("No account"));
+
+  req.unsavedAccount = req.session.unsavedAccount;
+  next();
+};
