+++ app/clients/dropbox/routes/moveExistingFiles.js
@@ -75,9 +75,8 @@ function DetermineFolder(title, client, callback) {
 }
 
 function Move(client, entries, callback) {
-
   if (!entries.length) return callback(null);
-  
+
   client
     .filesMoveBatch({
       entries: entries,
